(chore) Docker Compose can be used on CI to run the tests
* It would have been ideal to reuse docker-compose.test.yml as that would be exactly the same as we run in development *however* there is a critical issue with doing so which has lead to a copy being made. The issue is the following line from the docker-compose.test.yml:
```
volumes:
      - .:/srv/report-a-defect:cached
```
This line says ignore the files built into the image and overwrite them with whatever I is in the local directory. This is great for development as we can change files and they will be automatically picked up without a rebuilt. On CI there is no ability to mount volumes so it affectively says throw away the application and mount it to nothing.