(feature) Improve dashboard design
Added more Hackney styles for buttons and top header that live as components.
Moved estates section into a partial and used a secondary button to adhere to good practices of a single main call to action for a primary button which is a ‘Search’ in this case.

Changed button content from find to ‘search’.