(fix) Retain shellwords correctly when invoking rspec
The bin/dspec script uses the bash variable $@ to forward all its
arguments to rspec. However, if run like this:

    ./bin/dspec spec/features -e 'name of test'

The final shell argument `name of test` will be interpolated into the
rspec command like this, losing the quotation:

    docker exec ... rspec spec/features -e name of test

Quoting the "$@" argument means the shell will retain the arguments as
distinct shellwords, effectively keeping the quotation:

    docker exec ... rspec spec/features -e 'name of test'

More information can be found in the `Special Parameters` section of
`man bash`.