(feature) Record changes to a defect's status
In order to display changes to the status in the activity log, we need
to capture changes to them as part of the `activities` data that's
attached to the `Defect` model. An activity is created each time the
defect is updated, but it doesn't record any of the details of this
change.

We'd like to keep the recording of these changes as forward-compatible
as possible; in general, recording events like "changed column_a from
value_1 to value_2" risks encoding the current state of the schema into
long-term retained data, creating a coupling and long-term maintenance
cost.

For example, in this case, recording the before/after value of the
`status` field encodes the presence of the `status` field in perpetuity.
We've decided that's fine as it's a stable concept, unlikely to
radically change.

We're less comfortable with storing the magic numbers that Rails backs
this enum with -- if the set of statuses is changed in future,
maintainers will need to migrate any numbers stored in the event log,
which may not be apparent to them. They'd also need to version the data
so that it can still be interpreted, as a value of "3" may mean
something else in a year's time than it does today.

Therefore we're storing the status names in the event log, e.g.
`outstanding` or `follow_on`. These are symbolic and still require
conversion into UI strings, but they carry more intrinsic information
than a magic number would do, so we feel this is more future-proof.

Rails populates the `changes` hash when a model is mutated in memory,
but it is emptied before the `defect.update` activity is recorded. That
means we can't use it in the `params` lambda in the `tracked`
declaration. Instead we need to capture the status change in
`@tracked_changes` before the model is saved, and then use this value in
the `tracked` `params` hook to record the status change.

For the coupling reasons mentioned above, we're being conservative about
what we record and only storing the `status` change, not changes to
other fields.