(feature) Display status changes in the activity log
The Defect model can now generate activities with the key
`defect.update`, whose parameters may include a change to the status,
for example

    'status' => ['outstanding', 'follow_on']

Here we add logic to the DefectEventPresenter that we just extracted, to
handle such events.

This requires the status names to be converted to UI strings for
display, so I've exposed the Defect model's logic for doing this as a
class method.