(feature) Add sequence numbers to defects
To support having sequential reference numbers for defects, we need the
database table to have a SERIAL column that generates an
auto-incrementing unique number. This will generate a new value for each
row inserted.

The counter backing this column will also increment if an INSERT is
attempted but fails. This can happen when:

- a non-null column has not been given a value
- the new row violates a unique index constraint
- the INSERT is part of a transaction that ends up failing

With the constraints currently on the table, any non-null fields are
being populated by Rails or the app, so failed inserts should be very
rare and in practice the user will only end up seeing skips in the ID
sequence if any defect records are deleted.

We could in theory make a counter that was guaranteed not to skip values
on failed inserts, but that's a substantial amount more work and it does
not seem worth it given the very minor risk and effect of skipped IDs
happening.