(feature) Let defects.reference_number be null
As we're no longer using this column to provide the value of
`Defect#reference_number`, it does not need to be populated. To deploy
safely against the current version of the app, which is writing to this
column, and upgrade to a new release that no longer writes to it, we
allow it to be null rather than removing it.