(feature) Stop using the defects.reference_number column
Now that `defects.reference_number` can be null, and
`Defect#reference_number` does not use this column, we can stop
populating it when new records are created.

This should be deployed before the column is deleted, so that the
current running version is not affected by the column being removed.