(refactor) Extract ReferenceNumber presenter
In order to let staff search by reference number, we need a way to
determine whether their input is a reference number rather than a free
text input, and dispatch the query accordingly. Executing the query then
requires extracting the numeric data from the reference number format,
to send it to the database.

Here we extract a class that's responsible for generating the
`NB-xxx-xxx` display format for reference numbers, and can also
recognise this format and convert it back into a number. This keeps the
definition of the format in a single class, making it easier to adjust.