(feature) Search for defects by reference number
If a member of staff enters a reference number into the search field,
then we take them directly to that defect, rather than a search result
listing.

If they enter something that looks like a reference number but no such
defect is found, then we go back to the dashboard with a notice
containing an appropriate error message.