(feature) Accept reference numbers with non-standard formatting
Although reference numbers are canonically formatted as, for example,
`NB-123-456`, the staff may not necessarily type them in that way. For
example if a resident is reading their number over the phone, they may
omit the dashes.

This change means we'll recognise a search query as a reference number
if it contains the string "NB" followed by any number of digits and
dashes.