(fix) Patch security vulnerability in Omniauth
Use this gem https://github.com/cookpad/omniauth-rails_csrf_protection which is the advertised workaround as there is no official fix in the core gem available https://github.com/LBHackney-IT/report-a-defect/network/alert/Gemfile.lock/omniauth/open:
```
CVE-2015-9284 More information
high severity
Vulnerable versions: <= 1.9.0
Patched version: No fix
The request phase of the OmniAuth Ruby gem is vulnerable to Cross-Site Request Forgery when used as part of the Ruby on Rails framework, allowing accounts to be connected without user intent, user interaction, or feedback to the user. This permits a secondary account to be able to sign into the web application as the primary account.

In order to mitigate this vulnerability, Rails users should consider using the omniauth-rails_csrf_protection gem. More info is available in this PR.
```