(feature) Edit the target completion date of a defect
This adds a set of fields to the 'Edit defect' form that lets the user
set the completion date explicitly, instead of setting it implicitly by
setting a priority.

If a priority is set, this still takes precedence. Otherwise, if a date
is set, that will be written to the defect.

This reuses the date input partial we just extracted from the report
search form.

I've also disabled the Naming/AccessorMethodName, otherwise Rubocop
complains that the Defect#set_completion_date method should not begin
with the word `set_`. I'm not sure this is useful feedback, so I've
switched this cop off.