(fix) Correct date filtering spec
This test was passing by accident; when the page renders all the defects
created in this spec, it shows a table like this:

    +---------------+---------+---------+---------+
    | Total defects |       3 |       1 |       4 |
    +---------------+---------+---------+---------+

The first assertion in this spec checks for the presence of the number
4, and the second assertion checks for the digit 3, the intention being
to check that the given date parameters filter out one of the defects.

However, the test mixes up a month and a day value, and does not submit
the form, so both assertions run against the initial page load and both
pass.

This commit makes the test more strict by checking the complete summary
information, and fixes the bugs so that it passes properly.