(refactor) Extract date input partial
I'd like to reuse the date input fields that we have on the report
search form, in the form to edit the completion date of a defect.

Here I'm extracting this HTML and changing its interface slightly -- the
partial takes the name of the form fields, and a single date object,
rather than a presenter exposing multiple fields. This requires a
corresponding change in the controller to process the restructured
fields.