(fix) Styling for the 'Flagged' label on defect page
The 'Flagged label' is currently being displayed as a button because I
appropriated the `govuk-button` class to make it look like the nearby
buttons/links. So, it has `cursor: pointer` and depresses when clicked.

This changes it to look like the label that appears on the search
listing, and not have those unwanted interaction cues.