All defect types use the same Staff::Defects::CompletionController
Since the need to request the actual completion date when a defect's
status changes to completed is the same for both communal area and
property defects, let's use the same controller for both.

There's already a `defect_path_for` helper method, so we're able to
redirect back to the right place afterwards.