Fixes mismatches between lists of trades
Solves https://dxw.zendesk.com/agent/tickets/10096

Previously, the trade names listed in the trades-by-category arrays were out of sync with the aggregated list of all trades.

Source for TRADES: https://trello.com/c/hO8pC4gH/19-data-get-trades-list-for-defects-and-types

Source for trades by category: https://trello.com/c/4v3uUoVd/155-add-categories-as-parent-level-to-trades-to-support-reporting

We are using the list in `Defect::TRADES` as the canonical list because it is the list used to create new defects.

There are no existing defects, open or closed, using the mistyped strings.