Adds notion of added_at against a defect
We eventually want to have `created_at` by editable by a user of the
system. However we also want to maintain when something was added
to the system.

This commit adds a new attribute `added_at` which will default to
today's date (like `created_at`).

A migration was created to update all existing records to set
`added_at` to be the same as `created_at` and a model test has been
written to verify that new Defects get an `added_at`.