Adds `added_at` to CSV download
Adds a new header `date_added` to the CSV download.

Internally we use the Rails convention of `blah_at` but during
Sprint planning it was raised that this is a confusing convention
on the CSV.

For this reason we've went with `date_created` and renamed `created_at`
to be `date_created`.