Displays 'Created On' in tables and views
A user need was discovered that we want to display the created on
date pretty much anywhere in the system we display other dates

Another user need (to be tackled in this body of work, but not this
commit) is to allow users to edit created at as a date.

As a result this commit also changes how we display created at in places
we were already sharing it.

An unrelated change was to silently allow nil to be used as an owner name
in the case there is no owner for a change. This only appears to happen
when we first seed. This is definitely something we could improve. Perhaps
by delegating defect.owner.name in the Defect model.