Adds test for CombinedReportPresenter
There is some overlap between this Presenter and the SchemeReportPresenter.

For the purposes of testing I've split treated as two separate presenters,
this is because after speaking with the core user of the reports, it feels
like these reports are going to change and diverge in future iterations.

Tying them in too tightly together is probably a mistake.