Add Activity to seeded Defects
They were previously created with just a `defect.create` key, even if
the Defect was closed/completed.

All Defects are now created with a `defect.update` activity that assumes
they were created as "outstanding" and then updated to their current
status. Updating the activity more realistically (i.e. creating them
as "outstanding" and _then_ updating their status) is tricky, as it doesn't
assign an `owner` to the activity, so we're essentially lying about
their status here, but it's slightly closer to what actually happens.

Modify defect factory to generate Defects with `target_completion_date`s
in the past to generate some overdue ones.