Fix "overdue" behaviour
As far as I understand it, an "overdue" defect can be one of two things:
1. A defect that was completed _after_ its `target_completion_date`, but
that _was_ completed at some point during the past. This is an important
distinction for viewing past reports.
2. A defect that is still open after its `target_completion_date`. This
will be important for users viewing reports for the current/a recent
timeframe.
3. Some older defects may not have an actual completion date, as code to
add this date was written after they had been created and no migration
performed.  Staff can manually add this in by modifying the defect, so
I'm classing completed defects without an `actual_completion_date` as
overdue to encourage them to add the date when they've seen that the
numbers don't add up on their data export.

Before this change, _everything_ was being marked as "overdue" in the
past, as it was comparing the `target_completion_date` against the
current date, whether or not the defect was closed.