Fix "overdue" behaviour
As far as I understand it, an "overdue" defect can be one of two things:
1. A defect that was completed _after_ its `target_completion_date`, but
that _was_ completed at some point during the past. This is an important
distinction for viewing past reports.
2. A defect that is still open after its `target_completion_date`. This
will be important for users viewing reports for the current/a recent
timeframe.

Before this change, _everything_ was being marked as "overdue" in the
past, as it was comparing the `target_completion_date` against the
current date, whether or not the defect was closed.