WIP Fix combined report view
- Increase accuracy of defects completed on time

The percentage of defects completed in time should be a percentage
against the total number of completed defects _and_ those that are still
open and overdue.

Previously, we checked whether a report was completed on time based on
the last Activity record when a defect was updated with a "completed" status,
but this is unreliable due to the ordering of activities, and it makes more
sense to check against the defect's `actual_completion_date`, as this is more
reliable.

- Classify anything that is "closed" as completed.

Numbers on the previous report didn't match up with Hackney's own reporting
from CSV data export, as they included anything "closed" (e.g. rejected, raised
in error, completed, closed) as "completed". This now means numbers tally up
more accurately with their own records.

Hack: To add an `actual_completion_date` to most closed defects that are not
"completed", staff have to first mark it as "completed", then manually change
the status themselves to whatever it should be. This is obviously a pain, but
until we work on a fix for this, anything without an `actual_completion_date`
will be marked as overdue, as we have no way if it's been completed on time or
not.  This will hopefully encourage staff to double check whether everything
has an `actual_completion_date`.

TODO: FIX INCREDIBLE FLAKE IN TESTS