Improve accuracy of defects completed on time
The percentage of defects completed in time should be a percentage
against the total number of completed defects _and_ those that are still
open and overdue.

Previously, we checked whether a report was completed on time based on
the last Activity record when a defect was updated with a "completed"
status, but this is unreliable due to the ordering of activities, and it
makes more sense to check against the defect's `actual_completion_date`,
as this is more reliable.