Scheme and combined reports fixed by using the new added_at field
Since this code was written a new field called `added_at` was added to allow the users to overwrite the time a defect was added to the service. If they received it on the last day of the month but could only add it to the service in the second month, they needed a way to set the date back.

The reporting function did not keep up with this change. Defects that had their added_at field changed from the created_at started to be misreported.

This change moves the focus of the reports away from created_at and to the more reliable added_at field.