Fix install of Docker on CircleCI
When installing Docker there was a fatal error since it was no longer a supported version.

On reading how CircleCI advises Docker to be installed these days I found they have the concept of a convenience image[1], where Docker and Docker Compose are preinstalled.

Whilst this is a Ruby project, CI will be running everything through Docker so the 'base' image was chosen from the list[2].

[1] https://circleci.com/docs/2.0/docker-composec
[2] https://circleci.com/blog/announcing-our-next-generation-convenience-images-smaller-faster-more-deterministic/